Large code chunk distribution
Quite a handful of big changes. With out new Logistics class, we had to move some code from Utilities to this new class. Also there is a lot of variable renaming, code cleaning, and a new way to access shared data: through a single Logistics instance.
A new empty method was declared: packageDelivery, it'll be the next thing to implement.